feat: 정산 상세 조회 [S13P21A310-62]

See merge request s13-fintech-finance-sub1/S13P21A310!35

+++ backend/business/src/main/java/com/udong/backend/auth/repository/RefreshTokenRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByUser_Id(Long userId);          // 유저당 1행
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
+    Optional<RefreshToken> findByUser_Id(int userId);          // 유저당 1행
     Optional<RefreshToken> findByRefreshTokenHash(String hash); // 필요시 검증용
     long deleteByRefreshTokenHash(String refreshTokenHash);
 }
