+++ backend/business/src/main/java/com/udong/backend/users/repository/UserAvailabilityRepository.java
@@ -3,4 +3,4 @@ package com.udong.backend.users.repository;
 import com.udong.backend.users.entity.UserAvailability;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAvailabilityRepository extends JpaRepository<UserAvailability, Long> {}
+public interface UserAvailabilityRepository extends JpaRepository<UserAvailability, Integer> {}
